+++ src/Reflex/Host/Headless.hs
@@ -8,6 +8,7 @@ module Reflex.Host.Headless where
 
 import Control.Concurrent.Chan (newChan, readChan)
 import Control.Monad (unless)
+import Control.Monad.Catch (MonadCatch, MonadMask, MonadThrow)
 import Control.Monad.Fix (MonadFix, fix)
 import Control.Monad.IO.Class (MonadIO, liftIO)
 import Control.Monad.Primitive (PrimMonad)
@@ -23,26 +24,29 @@ import Reflex
 import Reflex.Host.Class
 
 type MonadHeadlessApp t m =
-  ( Adjustable t m
+  ( Reflex t
+  , Adjustable t m
+  , MonadCatch m
+  , MonadFix (Performable m)
   , MonadFix m
+  , MonadHold t (Performable m)
   , MonadHold t m
   , MonadIO (HostFrame t)
   , MonadIO (Performable m)
   , MonadIO m
+  , MonadMask m
   , MonadRef (HostFrame t)
+  , MonadSample t (Performable m)
+  , MonadSample t m
+  , MonadThrow m
   , NotReady t m
   , PerformEvent t m
   , PostBuild t m
   , PrimMonad (HostFrame t)
   , Ref (HostFrame t) ~ IORef
   , Ref m ~ IORef
-  , Reflex t
   , ReflexHost t
   , TriggerEvent t m
-  , MonadSample t (Performable m)
-  , MonadSample t m
-  , MonadFix (Performable m)
-  , MonadHold t (Performable m)
   )
 
 -- | Run a headless FRP network. Inside the action, you will most probably use